fix: 유저 견적서 css 수정
