+++ frontend/package.json
@@ -26,6 +26,7 @@
     "react-dom": "^18.3.1",
     "react-dropzone": "^14.2.3",
     "react-router-dom": "^6.26.1",
+    "react-zoom-pan-pinch": "^3.6.1",
     "swiper": "^11.1.10"
   },
   "devDependencies": {
