+++ frontend/src/styles/quotationRequest-list.module.scss
@@ -47,10 +47,10 @@
       & > .qr-card-grid {
 
         & > .qr-card {
+          width: 100%;
 
           & > .qr-card-content {
             margin: auto;
-
 
             & > .image-div {
               display: flex;
